nlp-service/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
pydantic==2.7.4
pytest==8.3.2
httpx==0.27.2
transformers==4.30.2
torch==2.7.1
sentence-transformers==2.2.2
aiohttp==3.8.4
tenacity==9.0.0
python-dotenv==1.0.1